ai-services/damage-detection/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
requests==2.32.3
python-dotenv==1.2.2
ultralytics==8.4.37
